Edited the filter.py with necessary changes

+++ Grayscale Image Filter/filter.py
@@ -1,6 +1,14 @@
+'''
+Install opencv and numpy before running this script:
+  
+pip install opencv-python numpy
+
+'''
+
 import urllib.request
 import cv2
 import numpy as np
+import os.path
 
 
 def grayscale_filter(image):
@@ -20,16 +28,17 @@ def grayscale_filter(image):
 
 
 # Path for the image
-image_path = './cat.png'
+image_path = './image.png'
 img = None
 
-try:
+if os.path.isfile(image_path):
     img = cv2.imread(image_path)
-except Exception:
+
+else:
     urllib.request.urlretrieve(
         'https://images.unsplash.com/photo-1615789591457-74a63395c990?ixid=MnwxMjA3fDB8MHxzZWFyY2h8MXx8YmFieSUyMGNhdHxlbnwwfHwwfHw%3D&ixlib=rb-1.2.1&w=1000&q=80',
-        "cat.png")
-    img = cv2.imread('cat.png')
+        "image.png")
+    img = cv2.imread('image.png')
 
 # convert to grayscale
 grayscaled_image = grayscale_filter(img)
